See #9. Try & failed to get logout to work as expected.
See http://stackoverflow.com/a/27809782/2578205

Once your login with a Google Account,
it pairs it with Google Chrome and
creates a Profile for it!
Then when you logout, it doesn't.
So essentially someone hits the "login" button and it logs in as the admin,
without prompting for password.
that is the security provided.